streamlit/requirements.txt
matplotlib==3.10.1
mysql_connector_repackaged==0.3.1
mysql-connector-python==9.2.0
numpy==2.2.4
pandas==2.2.3
prettytable==3.16.0
streamlit==1.44.0
